Check module minSDK against current SDK when loading.
Prior to checking module dependencies, each module will be checked
for the minimum SDK version it requires according to its manifest
and will be removed if it requests an SDK above the current version.

The target SDK version is not checked yet.

Include test for SDKVersion checking functionality in ModuleManager.

Closes #6