Refactoring ModuleManager from an RMI-based implementation to a Frame stack.
The previous RMI (remote method invocation) implementation of the modulemanager andassociated classes was fragile (it depended on heavy cross-process synchronization) and difficult to debug (because there were multiple processes involved in running a single module).

To deal with this, I have begun a rewrite of the ModuleManagerExecutor family of classes to instead use a stack of JFrames which are allocated on the fly as requested by modules.

At present, this approach works up to the point of closing out of a module; at this point in time the entire sdk crashes to desktop. I suspect there is an issue with object instances not being properly shared across threads, or maybe a race condition at the time of module shutdown. Unfortunately, my schedule during the semester did not leave me time to investigate what the problem might be.